pyproject.toml
[tool.poetry]
name = "pdf-investor-summarizer"
version = "0.1.0"
description = "LLM-powered extractor of investor-relevant data from PDF reports"
authors = ["Artem Shishkin <verter25@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pdfminer-six = "^20250506"
pdf2image = "^1.17.0"
pytesseract = "^0.3.13"
pillow = "^11.3.0"
langchain = "^0.3.26"
openai = "^1.93.0"
tiktoken = "^0.9.0"
pydantic = "^2.11.7"
typer = "^0.16.0"
requests = "^2.32.4"
pypdf = "^5.7.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-mock = "^3.14.1"
pytest-asyncio = "^1.0.0"
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.3.0"
mypy = "^1.16.1"
pre-commit = "^4.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
